code-challenge-v2/requirements.txt
boto3==1.42.60
dj-database-url==3.1.2
django-storages==1.14.6
django-webpack-loader==3.2.3
gunicorn==25.1.0
pip==25.3
psycopg2-binary==2.9.11
pytest==9.0.2
pytest-django==4.12.0
sentry-sdk==2.54.0
setuptools==82.0.0
whitenoise==6.12.0
djangorestframework==3.16.1
